I moved message_log.py to core and fixed the bugs related to it
