+++ src/main.py
@@ -20,7 +20,7 @@ from sv.ai import decide_enemy_action
 from sv.core.collision import MoveResult
 from sv.ui import ProgressBar
 
-from sv.ui.message_log import MessageLog
+from sv.core.message_log import MessageLog
 
 TILE_SIZE = Settings.TILE_SIZE
 PLAYER_INPUT_DIAGONAL_WINDOW = 0.02
