+++ src/sv/core/message_log.py
@@ -2,6 +2,7 @@ import time
 from collections import deque
 import arcade
 
+
 class Message:
     def __init__(self, text: str, kind: str):
         self.text = text
@@ -16,14 +17,7 @@ class MessageLog:
         "system": arcade.color.GOLD,
     }
 
-    def __init__(
-        self,
-        x=10,
-        y=100,
-        width=400,
-        line_height=18,
-        max_messages=8,
-    ):
+    def __init__(self, x=10, y=70, width=350, line_height=18, max_messages=6):
         self.x = x
         self.y = y
         self.width = width
@@ -31,32 +25,62 @@ class MessageLog:
         self.max_messages = max_messages
 
         self.messages: deque[Message] = deque()
+        self._text_objects: list[arcade.Text] = []
+        self._dirty = True
 
-    # API
+        # API
+
     def push(self, text: str, kind: str = "info"):
         self.messages.append(Message(text, kind))
 
-        # ограничение размера
         while len(self.messages) > self.max_messages:
             self.messages.popleft()
 
+        self._dirty = True
+
     def clear(self):
         self.messages.clear()
+        self._dirty = True
 
-    def draw(self):
-        """Рисует лог поверх HUD"""
+    # INTERNAL
+
+    def _rebuild(self):
+        self._text_objects.clear()
+
         y_offset = 0
 
-        for msg in self.messages:
+        for msg in reversed(self.messages):
             color = self.COLORS.get(msg.kind, arcade.color.WHITE)
 
-            arcade.draw_text(
+            text = arcade.Text(
                 msg.text,
                 self.x,
                 self.y + y_offset,
                 color,
                 14,
                 width=self.width,
+                multiline=True,
             )
 
+            self._text_objects.append(text)
             y_offset += self.line_height
+
+        self._dirty = False
+
+    #  DRAW
+
+    def draw(self):
+        # фон
+        arcade.draw_lrbt_rectangle_filled(
+            self.x - 5,
+            self.x + self.width,
+            self.y - 5,
+            self.y + self.line_height * self.max_messages + 5,
+            (0, 0, 0, 120),
+        )
+
+        if self._dirty:
+            self._rebuild()
+
+        for text in self._text_objects:
+            text.draw()
